+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change log
 
+## Version 0.10.0
+* Added typeahead convenience method
+* Added warning when invoking | operator on .search method
+* Added debug log info to execute_sql statement
+* Added logging to unit tests
+* Fixed a bug that caused the usage of # in a query resulting in a SQL parse error
+* When a NEAR operator is found in the query, the query is quoted.
+
 ## Version 0.9.0
 * Introduced Q objects for AND/OR queries
 * Introduced initial token queries using the allow_initial_token (^) lookup.